Update rooms list UI at set intervals
Previously, the rooms list tableView was refreshed every time a server replied with a list of rooms. Since the UI refresh can take several frames, this patch changes the refresh to only happen at 1, 2, and 3 seconds after sending the room requests, and to skip the refresh if no rooms were added.